move the group model definition in a specific model file
