+++ src/app/groups/containers/manager-permission-dialog/manager-permission-dialog.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
-import { Group } from '../../data-access/get-group-by-id.service';
+import { Group } from '../../models/group';
 import { Manager } from '../../data-access/get-group-managers.service';
 import { ProgressSelectValue, ProgressSelectComponent } from
   'src/app/ui-components/collapsible-section/progress-select/progress-select.component';
